feat: Add blog assistant functionality to llamabot
This commit introduces a new feature to the llamabot CLI, a blog assistant. The blog assistant can summarize and tag a blog post. The changes include:

- Importing the new 'blog' module in the CLI's __init__.py file.
- Adding a new 'blog' command to the CLI.
- Creating a new 'blog.py' file in the CLI directory.
- Creating a new 'blog.py' file in the prompt_library directory.
- Creating a new 'output_formatter.py' file in the prompt_library directory.
- Updating the 'blogging_assistant.ipynb' notebook to use the new 'coerce_dict' function from the 'output_formatter' module.

This new feature enhances the functionality of the llamabot CLI by providing users with a tool to automatically summarize and tag their blog posts.