feat: Extend llamabot installation with all optional dependencies
This commit modifies the GitHub workflow and the pyproject.toml file to install all optional dependencies of the llamabot package. This change will ensure that all features of the package are available during testing. 

In the GitHub workflow, the pipx install command for llamabot has been updated to include the [all] option. In the pyproject.toml file, a new section has been added to define the [all] option, which includes the tests, docs, and prompts dependencies. 

This change will help to improve the thoroughness of the testing process and ensure that all features of the package are working as expected.