test(python_prompt_library): disable deadline for ghostwriter test
This commit disables the deadline for the ghostwriter test in the Python prompt library. This is done to prevent Hypothesis from failing the test due to it taking too long to run. The change is made by adding the @settings decorator from Hypothesis with the deadline parameter set to None.