feat: add environment variable for model name
Added the ability to specify the model name via an environment variable.
This allows for more flexibility when deploying the bot in different environments.
The default model name is still "gpt-3.5-turbo-16k-0613" if the environment variable is not set.

Also, added the 'os' module to the import statements to facilitate the retrieval of the environment variable.

This change does not break any existing functionality.