feat: Add new Jupyter notebook for generating release notes
This commit introduces a new Jupyter notebook 'release-notes.ipynb' in the 'scratch_notebooks' directory.
The notebook contains code for generating release notes from git commit logs.
It uses the GitPython package to interact with the git repository and fetch commit logs.
It also includes a chatbot for answering queries related to GitPython usage.
The notebook is intended to help in automating the process of writing release notes for new versions of the software.