feat(configure): allow setting default model by name
This commit introduces a new feature in the `configure.py` file of the llamabot CLI.
The `default_model` function now accepts an optional `model_name` parameter.
If a valid model name is provided, it will be set as the default model.
If no model name is provided, or if the provided model name is not accessible with the user's API key, the user will be prompted to input a model name.

This change enhances the user experience by providing more flexibility in setting the default model.