feat: Add new prompt decorator and tests
This commit introduces a new prompt decorator in the scratch_notebooks directory, along with corresponding tests in the tests directory.
The new decorator is used to enhance the functionality of functions by adding a prompt feature.
The tests ensure that the decorator works as expected with different types of function arguments.

Additionally, this commit also includes tests for blogging prompts in the prompt_library directory.
These tests validate the output of different blogging prompt functions.