feat(querybot): Add support for response_tokens and history_tokens
This commit adds support for the `response_tokens` and `history_tokens` parameters in the `QueryBot` class.
These parameters allow the user to specify the number of tokens to use for the response and history in the chatbot.
The default values for these parameters are set to 2000 tokens each.

The `response_tokens` parameter determines the number of tokens to use for the chatbot's response,
while the `history_tokens` parameter determines the number of tokens to use for the chat history.
These parameters are used to calculate the token budget for the chatbot.

Additionally, the `chunk_sizes` parameter has been added to the `make_or_load_vector_index` function.
This parameter allows the user to specify a list of chunk sizes to use for the LlamaIndex TokenTextSplitter.
The default values for the `chunk_sizes` parameter are set to [200, 500, 1000, 2000, 5000].

These changes provide more flexibility and control over the token budget and chunk sizes used in the chatbot.