docs: Add instructions for using local Ollama models in LlamaBot
This commit adds instructions to the documentation on how to use local
Ollama models with LlamaBot. It includes a Python code snippet
demonstrating how to specify the `model_name` keyword argument when
creating a `SimpleBot` instance. It also mentions that the same can be
done for the `ChatBot` and `QueryBot` classes. A link to the Ollama
library of models is provided for reference.