feat: Add new Jupyter notebook for panel chat
This commit introduces a new Jupyter notebook 'panel-chat.ipynb' in the
'scratch_notebooks' directory. The notebook includes code for setting up
a chat interface using the Panel library, and integrating it with a
chatbot for interactive responses. It also includes functionality for
uploading a PDF file and having the bot summarize the content of the
file.