refactor(doc_processor): change return type of magic_load_doc function
The magic_load_doc function in doc_processor.py has been refactored.
The return type has been changed from a list of Document objects to a single Document object.
This change simplifies the handling of the returned document in the calling code.
The function now also initializes a Document object at the start of the function, which is then populated with the loaded document text.