refactor: move add_documents method to DocumentStore class
This commit moves the add_documents method from the QueryBot class to the DocumentStore class.
This change improves the organization of the code by placing the method in a more appropriate class.
The method is responsible for adding documents to the DocumentStore,
so it makes more sense for it to be a part of the DocumentStore class.
This change should not affect the functionality of the code.