chore(prompt_library): update model name to mistral-medium
The previous model name "gpt-4-1106-preview" has been replaced with "mistral/mistral-medium" in the `commitbot()` function of the `git.py` file. This change is a chore as it does not introduce a new feature or fix a bug, but rather updates the model used by the bot. This change should not have any impact on the functionality of the bot.

Note: This change does not include a BREAKING CHANGE as it does not modify the API or introduce any backwards-incompatible changes.