feat: add rank-bm25 library for BM25 ranking
This commit adds the `rank-bm25` library as a dependency to the project. This library provides an implementation of the BM25 ranking algorithm, which can be used to rank documents based on their relevance to a given query.

This new dependency will allow us to improve the accuracy of our search results by using a more sophisticated ranking algorithm.

The `rank-bm25` library is compatible with Python 3.10 and later, so there are no changes to the `requires-python` field in `pyproject.toml`.

This commit does not introduce any breaking changes.