feat(chatui): add support for initial message and serving the chat interface
This commit adds two new features to the `ChatUIMixin` class:

1. The ability to set an initial message that is displayed when the chat interface is loaded. This is done by adding a new `initial_message` parameter to the `__init__` method.
2. The ability to serve the chat interface directly, without having to manually call the `servable()` method and then show it. This is done by adding a new `serve` method to the class.

These changes should make it easier to use the `ChatUIMixin` class to create chat interfaces that have an initial message and can be served directly.

Note that these changes do not introduce any breaking changes, and should be fully backwards compatible with existing code that uses the `ChatUIMixin` class.