feat: add ChatUIMixin to QueryBot and update initialization and usage
The QueryBot class now includes the ChatUIMixin, which allows for more flexible handling of user input and output. The initialization of QueryBot has been updated to include an optional `initial_message` parameter and a `stream_target` parameter. The `__call__` method has also been updated to handle different stream targets.

The default number of results returned by the `__call__` method has been changed from 20 to 10.

Additionally, the import of `APIMixin` has been removed, as it is no longer used by QueryBot. The import of `ChatUIMixin` has been added in its place.

This change should be considered a breaking change, as it modifies the initialization and usage of QueryBot. Users of QueryBot will need to update their code to include the new parameters and handle the different stream targets.