feat: add support for different LLM models, initial message, and Panel integration
This commit adds several new features to the `chat` command:

- The ability to specify the name of the large language model (LLM) to use.
- The ability to specify an initial message for the bot.
- The ability to use Panel for the chat interface, instead of the default CLI chat.

The `model_name` option allows users to specify the name of the LLM to use. The default value is "mistral/mistral-medium".

The `initial_message` option allows users to specify the initial message that the bot should send to the user.

The `panel` option allows users to specify whether to use Panel for the chat interface. If this option is set to `True`, the chat interface will be served in a Panel. If it is set to `False`, the default CLI chat will be used.

The `stream_target` variable has been added to determine whether the chat interface should be served in the CLI or in a Panel. If `panel` is `True`, `stream_target` is set to "panel". Otherwise, it is set to "stdout".

The `bot` object has been updated to accept the new `model_name`, `initial_message`, and `stream_target` options.

Finally, if `panel` is `True`, the `bot.serve()` method is called to serve the chat interface in a Panel.